% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot for continuous variables}
\usage{
scatter_plot(data, var_x, var_y, na.rm = TRUE)
}
\arguments{
\item{data}{is the data frame whose columns you wish to plot}

\item{var_x}{is an integer (int) or double floating-point (dbl) value for the x-axis}

\item{var_y}{an integer (int) or double floating-point (dbl) value for the y-axis}

\item{na.rm}{logical argument, whether or not to remove the NAs, default setting is to remove the NAs (na.rm = TRUE)}
}
\value{
the scatter plot object as a ggplot
}
\description{
This function enables a quick way to generate a scatter plot for two continuous numerical variables in the dataset to check for any association between them
}
